Add `time_zone` and `env` args to `process_data()`
The `time_zone` argument applies the given time zone to date-times

The `env` argument specifies where the data should be loaded to. In most instances the function will be used interactively and the user will want to create the variables in the global environment, so this is the default. In testing or programatic applications it is useful to control where the variables are created, so an environment can be passed to capture them. Additionally, R CMD CHECK will no longer give a note for global assignment.